Always reset divert_update timer
With a grid_ie or solar feed that is updated at least once every second, the divert_update variable would never get a value other than zero, which caused its subscribers to not be updated when the timer is reset. We now define that subscribers should always be updated when the variable is assigned to, even if the new and old values are the same.